requirements.txt
# Flask 
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1

# database
PyMySQL==1.0.2
SQLAlchemy==1.4.23

#
python-dotenv==1.0.0
loguru==0.5.3
requests==2.26.0

# GUI
PySide6==6.2.0
numpy==1.21.2
matplotlib==3.4.3
koreanize-matplotlib==0.1.1

# 
pytest==6.2.5
black==21.9b0 